refactor(line_splitter): extract tests to sibling file (file-size gate)
split_lyrics_lines additions pushed line_splitter.rs to 1060 lines (>1000
CI cap). Move the #[cfg(test)] mod to line_splitter_tests.rs via #[path];
impl drops to ~200 lines. No code change.
