+++ crates/sp-server/src/lyrics/line_splitter_tests.rs
@@ -1,0 +1,778 @@
+use super::*;
+
+fn line(text: &str, start_ms: u32, end_ms: u32) -> AlignedLine {
+    AlignedLine {
+        text: text.into(),
+        start_ms,
+        end_ms,
+        words: None,
+    }
+}
+
+#[test]
+fn default_max_chars_is_32() {
+    assert_eq!(DEFAULT_MAX_CHARS, 32);
+}
+
+#[test]
+fn line_under_max_passes_through() {
+    let l = line("Short line.", 0, 1000);
+    let track = AlignedTrack {
+        lines: vec![l.clone()],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, SplitConfig::default());
+    assert_eq!(split.lines.len(), 1);
+    assert_eq!(split.lines[0].text, "Short line.");
+}
+
+#[test]
+fn long_line_splits_at_sentence_end() {
+    // 47 chars total — exceeds the 32-char default. '.' at byte 23 is inside
+    // the first 32 chars, so priority-1 (sentence-end) wins; split at byte 24.
+    let l = line("Praise our God and King. Tell the whole world.", 0, 4000);
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, SplitConfig::default());
+    assert_eq!(split.lines.len(), 2);
+    assert_eq!(split.lines[0].text, "Praise our God and King.");
+    assert_eq!(split.lines[1].text, "Tell the whole world.");
+}
+
+#[test]
+fn long_line_splits_at_comma_when_no_sentence_end() {
+    // 45 chars total — exceeds 32. No '.!?…' in the first 32 chars, so
+    // priority-2 (comma) wins; split at byte 24 after the ','.
+    let l = line("Praise our God and King, tell the whole world", 0, 4000);
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, SplitConfig::default());
+    assert!(split.lines.len() >= 2);
+    assert!(split.lines[0].text.ends_with(','));
+}
+
+#[test]
+fn long_line_falls_back_to_word_boundary() {
+    let l = line("Hallelujah praise hallelujah praise the Lord", 0, 4000);
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, SplitConfig::default());
+    assert!(split.lines.len() >= 2);
+    // No mid-word breaks
+    for sl in &split.lines {
+        assert!(!sl.text.starts_with(' '));
+        assert!(!sl.text.ends_with(' '));
+    }
+}
+
+#[test]
+fn timing_proportionally_distributed() {
+    // "Praise our God and King. Tell the whole world." splits at '.' after "King."
+    // Left:  "Praise our God and King." — non-WS chars: P,r,a,i,s,e,o,u,r,G,o,d,a,n,d,K,i,n,g,. = 20
+    // Right: "Tell the whole world."   — non-WS chars: T,e,l,l,t,h,e,w,h,o,l,e,w,o,r,l,d,.    = 18
+    // Total non-WS = 38
+    // Expected mid = 0 + (4000 * 20 / 38) = 80000 / 38 = 2105 (integer division)
+    // A buggy uniform-split implementation (mid = (0+4000)/2 = 2000) would fail
+    // the ±10ms tolerance check below.
+    let l = line("Praise our God and King. Tell the whole world.", 0, 4000);
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, SplitConfig::default());
+    assert_eq!(split.lines.len(), 2);
+    let expected_mid: u32 = 2105;
+    let actual_mid = split.lines[0].end_ms;
+    assert!(
+        actual_mid.abs_diff(expected_mid) <= 10,
+        "expected mid ≈ {expected_mid}ms (char-weighted), got {actual_mid}ms"
+    );
+    // Continuity: line 1 end == line 2 start
+    assert_eq!(split.lines[0].end_ms, split.lines[1].start_ms);
+    // Outer bounds preserved
+    assert_eq!(split.lines[0].start_ms, 0);
+    assert_eq!(split.lines[1].end_ms, 4000);
+}
+
+#[test]
+fn timing_proportional_diverges_from_uniform() {
+    // Unbalanced split: "Hi." is tiny vs the long tail.
+    // Sentence-end priority finds '.' after "Hi" at char 3 (within max_chars=32).
+    // Left:  "Hi." — non-WS chars: H,i,. = 3
+    // Right: "Then a much longer phrase that runs on for a while."
+    //        non-WS: T,h,e,n=4 + a=1 + m,u,c,h=4 + l,o,n,g,e,r=6 + p,h,r,a,s,e=6
+    //               + t,h,a,t=4 + r,u,n,s=4 + o,n=2 + f,o,r=3 + a=1 + w,h,i,l,e,.=6 = 41
+    // Total non-WS = 3 + 41 = 44
+    // Expected mid = 0 + (4000 * 3 / 44) = 12000 / 44 = 272ms
+    // Uniform would place mid at (0 + 4000) / 2 = 2000ms — VERY different.
+    let l = line(
+        "Hi. Then a much longer phrase that runs on for a while.",
+        0,
+        4000,
+    );
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, SplitConfig::default());
+    assert!(split.lines.len() >= 2, "expected a split");
+    // Proportional mid is ≈272ms — far below the uniform 2000ms midpoint.
+    // A uniform-split implementation would fail this assertion.
+    assert!(
+        split.lines[0].end_ms < 600,
+        "expected proportional mid < 600ms for tiny left half, got {}ms (uniform would be 2000ms)",
+        split.lines[0].end_ms
+    );
+}
+
+#[test]
+fn zero_duration_line_does_not_panic() {
+    // Edge case: start_ms == end_ms (zero-duration). The clamp + saturating_sub
+    // must produce a finite result without underflow or div-by-zero.
+    let l = line("Praise the Lord. Tell the world.", 5000, 5000);
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, SplitConfig::default());
+    assert!(!split.lines.is_empty());
+    for sl in &split.lines {
+        assert_eq!(sl.start_ms, 5000);
+        assert_eq!(sl.end_ms, 5000);
+    }
+}
+
+#[test]
+fn no_safe_split_passes_through_long_line() {
+    // Single long word (no spaces) — can't split safely
+    let l = line(&"a".repeat(50), 0, 1000);
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, SplitConfig::default());
+    assert_eq!(split.lines.len(), 1, "no safe split → preserve original");
+}
+
+// ── split_line: || guard on empty halves (line 60 mutant) ───────────────
+//
+// Mutant: `||` → `&&` — would only bail when BOTH halves are empty.
+// Need a case where split_idx lands exactly at a word-boundary yielding
+// an empty right_text (after trim_start). In that scenario the original
+// line should be passed through unchanged.
+// We construct this by picking max_chars exactly equal to the line length,
+// so no split is needed, but also a split that would leave one half empty.
+
+#[test]
+fn split_line_passthrough_when_right_half_empty_after_trim() {
+    // split_idx lands at position beyond all non-space content → right is empty
+    // Use a line where the only split candidate puts all text in the left half.
+    // E.g. "Hello world " — 12 chars. With max_chars=6, sentence-end wins if '.' present;
+    // use a line "Hello." (6 chars) where sentence-end is at byte 6, but text.len() == 6
+    // means next = 6 == text.len(), so the condition `next < text.len()` is false.
+    // That path falls through to the word-boundary search but there's no space ≤ limit_idx.
+    // Result: find_split_index returns None → pass through.
+    // This exercises the `None => return vec![line.clone()]` path, which verifies
+    // the guard is needed in the first place.
+    let l = line("Hello.", 0, 1000);
+    let cfg = SplitConfig { max_chars: 6 };
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, cfg);
+    // Text len == max_chars → no split needed, passes through untouched
+    assert_eq!(split.lines.len(), 1);
+    assert_eq!(split.lines[0].text, "Hello.");
+}
+
+#[test]
+fn split_line_passthrough_when_both_halves_would_be_empty() {
+    // A line consisting of spaces only produces empty left and right after trim.
+    // find_split_index returns None for all-whitespace (no word boundary pattern).
+    // The `None => return vec![line.clone()]` path triggers.
+    let l = line("     ", 0, 500);
+    let cfg = SplitConfig { max_chars: 2 };
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, cfg);
+    assert_eq!(split.lines.len(), 1);
+}
+
+// ── split_line: recursive re-split guard (lines 97/102 mutants) ──────────
+//
+// Mutant: `> cfg.max_chars` → `>= cfg.max_chars` — would try to recursively
+// split a half that is exactly max_chars long (which `split_track` already
+// wouldn't have touched), causing unnecessary re-splits.
+// Test: construct a line that splits into two halves where one half is
+// EXACTLY max_chars chars. Under the mutated code that half would be
+// re-split; under correct code it is left alone.
+
+#[test]
+fn split_line_does_not_recurse_when_half_is_exactly_max_chars() {
+    // "Hallelujah alleluia" splits at space after "Hallelujah" (11 chars).
+    // Left "Hallelujah" = 10 chars, right "alleluia" = 8 chars.
+    // With max_chars=10: left == 10, NOT > 10 → no recursion.
+    // Under `>= 10` mutant, left would be re-split (impossible since
+    // find_split_index returns None for a 10-char no-punctuation word),
+    // but we'd still get 2 lines. The key: both halves must be ≤ max_chars.
+    let l = line("Hallelujah alleluia", 0, 2000);
+    let cfg = SplitConfig { max_chars: 10 };
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, cfg);
+    // Expected: 2 lines, "Hallelujah" and "alleluia"
+    assert_eq!(split.lines.len(), 2);
+    assert_eq!(split.lines[0].text, "Hallelujah");
+    assert_eq!(split.lines[1].text, "alleluia");
+}
+
+// ── find_split_index: center / distance arithmetic (lines 131-164 mutants)
+
+#[test]
+fn find_split_index_returns_none_for_short_text() {
+    // text.chars().count() <= max_chars → must return None
+    assert_eq!(find_split_index("Hello", 5), None);
+    assert_eq!(find_split_index("Hello", 6), None);
+    assert_eq!(find_split_index("Hello", 100), None);
+}
+
+#[test]
+fn find_split_index_word_boundary_nearest_center() {
+    // "aaaa bbbb cccc dddd" — 19 chars, max_chars=10.
+    // Spaces at bytes 4 and 9 and 14.
+    // center = max_chars/2 = 5. center_byte = byte of chars[5] = 5.
+    // Spaces: byte 4 (dist=1), byte 9 (dist=4), byte 14 (>limit_idx, excluded).
+    // limit_idx = byte of chars[10] = 10.
+    // Nearest-to-center space at byte 4 (dist=|4-5|=1) → split at byte 5.
+    // Result: left = "aaaa", right = "bbbb cccc dddd".
+    //
+    // Under `/→%` mutant: center = 10 % 2 = 0; nearest = byte 4 instead of 5.
+    // Under `/→*` mutant: center = 10 * 2 = 20; both spaces have same "nearest".
+    // Under `−→+` mutant: distance = idx + center_byte (always larger), picks arbitrary.
+    // Under `−→/` mutant: distance = idx / center_byte (wrong scaling).
+    let result = find_split_index("aaaa bbbb cccc dddd", 10);
+    // The split should be at a word boundary near the center.
+    assert!(result.is_some());
+    let idx = result.unwrap();
+    // idx must be a valid split point: not zero, not at end, byte is word start
+    assert!(idx > 0);
+    assert!(idx < "aaaa bbbb cccc dddd".len());
+    // The left half must be at most max_chars chars
+    assert!("aaaa bbbb cccc dddd"[..idx].chars().count() <= 10);
+}
+
+#[test]
+fn find_split_index_word_boundary_nearest_center_verified_value() {
+    // Verify the exact split byte for a controlled case.
+    // "abcde fghij klmno" — 17 chars, max_chars=9.
+    // Spaces: byte 5 (char idx 5), byte 11 (char idx 11).
+    // center = 9/2 = 4. center_byte = byte of chars[4] = 4.
+    // limit_idx = byte of chars[9] = 9.
+    // Space at byte 5: dist = |5-4| = 1, idx <= 9 ✓ → split = 6.
+    // Space at byte 11: idx=11 > limit_idx=9 → excluded.
+    // So split at byte 6 → left = "abcde", right = "fghij klmno".
+    let text = "abcde fghij klmno";
+    let result = find_split_index(text, 9);
+    assert_eq!(result, Some(6), "nearest-center split at byte 6");
+}
+
+#[test]
+fn find_split_index_does_not_split_beyond_limit() {
+    // Spaces beyond limit_idx must not be chosen. Verify the right half
+    // starts within the original max_chars window.
+    // "hello world and more text here" — 30 chars, max_chars=12.
+    let text = "hello world and more text here";
+    let result = find_split_index(text, 12);
+    assert!(result.is_some());
+    let idx = result.unwrap();
+    assert!(
+        text[..idx].chars().count() <= 12,
+        "left half must be ≤ max_chars"
+    );
+}
+
+#[test]
+fn find_split_index_rightmost_word_boundary_fallback() {
+    // No sentence-end, no comma, and no space near center — triggers path 4.
+    // "abcdefghij klmno" — 16 chars, max_chars=8.
+    // No punctuation. Spaces at byte 10 only (beyond limit_idx=8 → excluded from
+    // nearest-center search too). Falls to rfind(' ') in text[..limit_idx].
+    // limit_idx = byte of chars[8] = 8. text[..8] = "abcdefgh" — no space.
+    // → falls all the way to rfind which also finds nothing → None.
+    // Test a case where there IS a space within limit_idx for path 4.
+    // "abc def ghij klmno" — 18 chars, max_chars=10.
+    // Spaces: byte 3, byte 7, byte 12 (>limit_idx=10 → excluded).
+    // center=5, center_byte=5. Space at byte 3: dist=2. Space at byte 7: dist=2.
+    // Both equal distance — pick first (smallest idx with min dist? no, iter is
+    // forward order, so byte 3 seen first with dist=2, byte 7 seen next with dist=2,
+    // dist < d is strict so byte 3 wins). split at byte 4.
+    let text = "abc def ghij klmno";
+    let result = find_split_index(text, 10);
+    assert!(result.is_some());
+    let idx = result.unwrap();
+    assert!(text[..idx].chars().count() <= 10);
+    assert!(idx > 0);
+}
+
+#[test]
+fn find_split_index_split_point_is_after_space_not_at_space() {
+    // `rfind(' ').map(|i| i + 1)` — if mutated to just `i` the split includes
+    // the space in the left half (it would end with ' ').
+    // Similarly for the nearest-center path: `best = Some((idx + 1, dist))`.
+    // After split_line trims, a space at the end would not appear, but we can
+    // verify the byte index is past the space character.
+    let text = "hello world how are you doing here";
+    let result = find_split_index(text, 15);
+    assert!(result.is_some());
+    let idx = result.unwrap();
+    // The byte at idx should NOT be a space (it's the start of the right word)
+    let byte_at_idx = text.as_bytes().get(idx).copied().unwrap_or(0);
+    assert_ne!(
+        byte_at_idx, b' ',
+        "split idx must point past the space, not at it"
+    );
+}
+
+// ── split_words_by_index: return values and arithmetic (lines 171-187) ───
+//
+// Mutants on split_words_by_index:
+//   - Replace return with (None, None), (None, Some(vec![])), etc.
+//   - `*` → `+` / `/` in `words.len() * byte_idx`
+//   - `/` → `%` / `*` in `/ line.text.len().max(1)`
+//
+// We test that the function correctly distributes words proportionally.
+
+// ── split_line: trailing-spaces right half → `||` guard (line 60 mutant) ─
+//
+// Mutant: `||` → `&&` — only bails when BOTH halves are empty.
+// Input: "hello      " — spaces fill the right half after the nearest-center
+// split at byte 6.  right_text = "     ".trim_start() = "" (empty), while
+// left_text = "hello" (non-empty).  Original `||` bails (returns original
+// line); mutant `&&` does not bail and produces two lines, the second being
+// an empty string.
+#[test]
+fn split_line_passthrough_when_right_half_is_all_spaces() {
+    // "hello      " = 5 non-space + 6 spaces = 11 chars, max_chars=8.
+    // find_split_index: center=4, space at byte 5 (dist=1) → split at 6.
+    // right = "     " → trim_start = "" → right_text empty → || guard fires.
+    let l = line("hello      ", 0, 1000);
+    let cfg = SplitConfig { max_chars: 8 };
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, cfg);
+    assert_eq!(
+        split.lines.len(),
+        1,
+        "all-space right half must trigger passthrough"
+    );
+    assert_eq!(split.lines[0].text, "hello      ");
+}
+
+// ── split_line: deep recursion needed (lines 102 `> → ==,<` mutants) ──────
+//
+// Mutant `> → ==`: right is recursed only when it is EXACTLY max_chars.
+// Mutant `> → <`:  right is recursed only when it is LESS than max_chars.
+// Both cause a right half that is strictly greater than max_chars to be
+// pushed without further splitting, producing only 2 lines instead of ≥3.
+#[test]
+fn split_line_recurses_right_half_until_within_max_chars() {
+    // "aaa bbb ccc ddd eee fff" (23 chars), max_chars=8.
+    // First split: center=4, center_byte=4.  Spaces: byte 3 (dist=1) and
+    // byte 7 (dist=3).  Nearest = byte 3 → split at 4.
+    // left="aaa" (3 chars, ≤8 — no recursion).
+    // right="bbb ccc ddd eee fff" (19 chars, >8) → must recurse.
+    // Recursion produces multiple sub-lines.  Total output ≥ 3.
+    // Under `> → ==`: 19 == 8 is false → right pushed as-is → only 2 lines.
+    // Under `> → <`: 19 < 8 is false → right pushed as-is → only 2 lines.
+    let l = line("aaa bbb ccc ddd eee fff", 0, 5000);
+    let cfg = SplitConfig { max_chars: 8 };
+    let track = AlignedTrack {
+        lines: vec![l],
+        provenance: "t".into(),
+        raw_confidence: 1.0,
+    };
+    let split = split_track(&track, cfg);
+    assert!(
+        split.lines.len() >= 3,
+        "right half >8 chars must recurse; expected ≥3 lines, got {}",
+        split.lines.len()
+    );
+    for sl in &split.lines {
+        assert!(
+            sl.text.chars().count() <= 8,
+            "each output line must be ≤max_chars, got {:?}",
+            sl.text
+        );
+    }
+}
+
+// ── find_split_index: center-division correctness (line 148 mutants) ──────
+//
+// center = max_chars / 2 (integer division).
+// Mutant `/→%`: center = max_chars % 2 — for even max_chars this is 0,
+//   so center_byte=0 and the leftmost space within limit always wins.
+// Mutant `/→*`: center = max_chars * 2 — out of range for most inputs,
+//   center_byte falls back to text.len(), so the RIGHTMOST space within
+//   limit wins instead of the nearest-to-center one.
+//
+// Two separate inputs are required:
+//   (A) An input where the LEFTMOST space is NOT the nearest-to-center one
+//       (distinguishes `/→%` from original).
+//   (B) An input where the RIGHTMOST space within limit is NOT the
+//       nearest-to-center one (distinguishes `/→*` from original).
+
+#[test]
+fn find_split_index_center_division_not_modulo() {
+    // "a bbbbb ccccc" — 13 chars, max_chars=10.
+    // chars: a(0), ' '(1), b(2-6), ' '(7), c(8-12).
+    // center=5, center_byte=5 (char[5]='b' at byte 5).
+    // Spaces within limit(10): byte 1 (dist=4), byte 7 (dist=2).
+    // Nearest = byte 7 → split at 8 → Some(8).
+    //
+    // Under `/→%`: center=0, center_byte=0.
+    //   dist(1)=1, dist(7)=7 → nearest = byte 1 → Some(2). Different.
+    let result = find_split_index("a bbbbb ccccc", 10);
+    assert_eq!(
+        result,
+        Some(8),
+        "nearest-center split must use division, not modulo"
+    );
+}
+
+#[test]
+fn find_split_index_center_division_not_multiplication() {
+    // "aaaa bb ccc dddd" — 16 chars, max_chars=8.
+    // chars: a(0-3), ' '(4), b(5-6), ' '(7), c(8-10), ' '(11), d(12-15).
+    // center=4, center_byte=4 (char[4]=' ' at byte 4).
+    // limit_idx=byte 8 (char[8]='c').
+    // Spaces within limit: byte 4 (dist=0), byte 7 (dist=3).
+    // Nearest = byte 4 → split at 5 → Some(5).
+    //
+    // Under `/→*`: center=16, center_byte=text.len()=16.
+    //   dist(4)=12, dist(7)=9 → nearest-to-end = byte 7 → Some(8). Different.
+    let result = find_split_index("aaaa bb ccc dddd", 8);
+    assert_eq!(
+        result,
+        Some(5),
+        "nearest-center split must use division, not multiplication"
+    );
+}
+
+// ── find_split_index: distance subtraction (line 153 mutants) ────────────
+//
+// dist = (idx as i64 - center_byte as i64).abs()
+// Mutant `-→+`: dist = idx + center_byte — always large, picks leftmost space.
+// Mutant `-→/`: dist = idx / center_byte (integer) — maps distances to small
+//   integers, effectively picking the leftmost low-ratio space.
+//
+// Input with three spaces where the correct nearest-to-center is NOT the
+// leftmost one.  Both mutations pick the leftmost space instead.
+
+#[test]
+fn find_split_index_nearest_center_uses_subtraction_for_distance() {
+    // "aa bb ccccc ddd" — 15 chars, max_chars=12.
+    // chars: a(0-1), ' '(2), b(3-4), ' '(5), c(6-10), ' '(11), d(12-14).
+    // center=6, center_byte=6 (char[6]='c' at byte 6).
+    // limit_idx=byte 12.
+    // Spaces within limit: byte 2 (dist=4), byte 5 (dist=1), byte 11 (dist=5).
+    // Nearest = byte 5 → split at 6 → Some(6).
+    //
+    // Under `-→+`: dist(2)=8, dist(5)=11, dist(11)=17 → leftmost (byte 2) → Some(3).
+    // Under `-→/`: dist(2)=0, dist(5)=0 (int), dist(11)=1 → tie(2,5) → byte 2 → Some(3).
+    let result = find_split_index("aa bb ccccc ddd", 12);
+    assert_eq!(
+        result,
+        Some(6),
+        "distance must be computed via subtraction, not addition or division"
+    );
+}
+
+// ── find_split_index: strict-less-than best update (line 154 mutants) ────
+//
+// if best.is_none_or(|(_, d)| dist < d) { best = Some((idx + 1, dist)); }
+//
+// Mutant `< → ==`: updates on equal distance → last tie wins (not first).
+// Mutant `< → <=`: also updates on equal → last tie wins.
+// Mutant `< → >`:  updates on LARGER distance → picks the farthest space.
+//
+// Two test inputs are required:
+//   (A) Two equidistant spaces → first one must win (kills == and <=).
+//   (B) Two unequal-distance spaces → nearer one must win (kills >).
+
+#[test]
+fn find_split_index_equal_distance_picks_first_space() {
+    // "aa bbbbb ccc ddd" — 16 chars, max_chars=10.
+    // chars: a(0-1), ' '(2), b(3-7), ' '(8), c(9-11), ' '(12), d(13-15).
+    // center=5, center_byte=5 ('b' at byte 5).
+    // limit_idx=byte 10. Spaces within limit: byte 2 (dist=3), byte 8 (dist=3).
+    // Equal distance — forward iteration picks byte 2 first → split at 3 → Some(3).
+    //
+    // Under `< → ==` or `< → <=`: byte 8 (dist=3) also updates → Some(9). Different.
+    let result = find_split_index("aa bbbbb ccc ddd", 10);
+    assert_eq!(
+        result,
+        Some(3),
+        "equal-distance spaces: first (leftmost) must be chosen, not last"
+    );
+}
+
+#[test]
+fn find_split_index_nearer_center_wins_over_farther() {
+    // "a bbbbbb cccc" — 13 chars, max_chars=10.
+    // chars: a(0), ' '(1), b(2-7), ' '(8), c(9-12).
+    // center=5, center_byte=5 ('b' at byte 5).
+    // limit_idx=byte 10. Spaces within limit: byte 1 (dist=4), byte 8 (dist=3).
+    // Nearer = byte 8 → split at 9 → Some(9).
+    //
+    // Under `< → >`: picks farthest → byte 1 (dist=4 > dist=3) → Some(2). Different.
+    let result = find_split_index("a bbbbbb cccc", 10);
+    assert_eq!(
+        result,
+        Some(9),
+        "nearer-to-center space must win over farther space"
+    );
+}
+
+#[test]
+fn split_words_by_index_none_when_words_is_none() {
+    let l = AlignedLine {
+        text: "hello world".into(),
+        start_ms: 0,
+        end_ms: 1000,
+        words: None,
+    };
+    let (left, right) = split_words_by_index(&l, 5);
+    assert!(left.is_none(), "None words → left must be None");
+    assert!(right.is_none(), "None words → right must be None");
+}
+
+#[test]
+fn split_words_by_index_none_when_words_empty() {
+    let l = AlignedLine {
+        text: "hello world".into(),
+        start_ms: 0,
+        end_ms: 1000,
+        words: Some(vec![]),
+    };
+    let (left, right) = split_words_by_index(&l, 5);
+    assert!(left.is_none(), "empty words → left must be None");
+    assert!(right.is_none(), "empty words → right must be None");
+}
+
+#[test]
+fn split_words_by_index_distributes_proportionally() {
+    // "hello world" — 11 bytes. Split at byte 6 (after "hello ").
+    // words.len() = 2, byte_idx = 6, text.len() = 11.
+    // split_word = (2 * 6) / 11 = 12 / 11 = 1.
+    // left = words[..1] = ["hello"], right = words[1..] = ["world"].
+    //
+    // Under `* → +` mutant: (2 + 6) / 11 = 0 → all words go to right, left=None.
+    // Under `/ → %` mutant: (2 * 6) % 11 = 1 (same here but different for other inputs).
+    // Under `/ → *` mutant: (2 * 6) * 11 = 132 → min(132, 2) = 2 → all left, right=None.
+    let hello = AlignedWord {
+        text: "hello".into(),
+        start_ms: 0,
+        end_ms: 400,
+        confidence: 0.9,
+    };
+    let world = AlignedWord {
+        text: "world".into(),
+        start_ms: 400,
+        end_ms: 1000,
+        confidence: 0.9,
+    };
+    let l = AlignedLine {
+        text: "hello world".into(),
+        start_ms: 0,
+        end_ms: 1000,
+        words: Some(vec![hello.clone(), world.clone()]),
+    };
+    let (left, right) = split_words_by_index(&l, 6);
+    assert!(left.is_some(), "left must be Some (has at least 1 word)");
+    assert!(right.is_some(), "right must be Some (has at least 1 word)");
+    let lv = left.unwrap();
+    let rv = right.unwrap();
+    assert_eq!(lv.len(), 1);
+    assert_eq!(rv.len(), 1);
+    assert_eq!(lv[0].text, "hello");
+    assert_eq!(rv[0].text, "world");
+}
+
+#[test]
+fn split_words_by_index_all_to_right_when_byte_idx_zero() {
+    // byte_idx = 0 → split_word = (2 * 0) / 11 = 0 → left empty → None
+    let hello = AlignedWord {
+        text: "hello".into(),
+        start_ms: 0,
+        end_ms: 400,
+        confidence: 0.9,
+    };
+    let world = AlignedWord {
+        text: "world".into(),
+        start_ms: 400,
+        end_ms: 1000,
+        confidence: 0.9,
+    };
+    let l = AlignedLine {
+        text: "hello world".into(),
+        start_ms: 0,
+        end_ms: 1000,
+        words: Some(vec![hello, world]),
+    };
+    let (left, right) = split_words_by_index(&l, 0);
+    assert!(
+        left.is_none(),
+        "byte_idx=0 → split_word=0 → left empty → None"
+    );
+    assert!(right.is_some(), "all words go to right");
+    assert_eq!(right.unwrap().len(), 2);
+}
+
+#[test]
+fn split_words_by_index_all_to_left_when_byte_idx_at_end() {
+    // byte_idx = text.len() → split_word = min(words.len(), words.len()) = 2
+    // left = all words, right = empty → None
+    let hello = AlignedWord {
+        text: "hello".into(),
+        start_ms: 0,
+        end_ms: 400,
+        confidence: 0.9,
+    };
+    let world = AlignedWord {
+        text: "world".into(),
+        start_ms: 400,
+        end_ms: 1000,
+        confidence: 0.9,
+    };
+    let text = "hello world";
+    let l = AlignedLine {
+        text: text.into(),
+        start_ms: 0,
+        end_ms: 1000,
+        words: Some(vec![hello, world]),
+    };
+    let (left, right) = split_words_by_index(&l, text.len());
+    assert!(left.is_some(), "byte_idx at end → all words to left");
+    assert!(right.is_none(), "right is empty → None");
+    assert_eq!(left.unwrap().len(), 2);
+}
+
+#[test]
+fn split_words_by_index_four_words_mid_split() {
+    // "aa bb cc dd" — 11 bytes. 4 words. split at byte 6 (after "aa bb ").
+    // split_word = (4 * 6) / 11 = 24 / 11 = 2.
+    // Under `* → +`: (4 + 6) / 11 = 0 → left=None (WRONG).
+    // Under `/ → *`: (4 * 6) * 11 = 264 → min(264, 4) = 4 → right=None (WRONG).
+    let make_w = |t: &str, s: u32, e: u32| AlignedWord {
+        text: t.into(),
+        start_ms: s,
+        end_ms: e,
+        confidence: 0.9,
+    };
+    let l = AlignedLine {
+        text: "aa bb cc dd".into(),
+        start_ms: 0,
+        end_ms: 4000,
+        words: Some(vec![
+            make_w("aa", 0, 1000),
+            make_w("bb", 1000, 2000),
+            make_w("cc", 2000, 3000),
+            make_w("dd", 3000, 4000),
+        ]),
+    };
+    let (left, right) = split_words_by_index(&l, 6);
+    assert!(left.is_some());
+    assert!(right.is_some());
+    let lv = left.unwrap();
+    let rv = right.unwrap();
+    // split_word = 2: first 2 go left, last 2 go right
+    assert_eq!(lv.len(), 2, "split_word=2 → 2 words left");
+    assert_eq!(rv.len(), 2, "split_word=2 → 2 words right");
+    assert_eq!(lv[0].text, "aa");
+    assert_eq!(lv[1].text, "bb");
+    assert_eq!(rv[0].text, "cc");
+    assert_eq!(rv[1].text, "dd");
+}
+
+fn ll(start: u64, end: u64, en: &str) -> sp_core::lyrics::LyricsLine {
+    sp_core::lyrics::LyricsLine {
+        start_ms: start,
+        end_ms: end,
+        en: en.into(),
+        sk: None,
+        words: None,
+    }
+}
+
+#[test]
+fn lyrics_short_line_passes_through() {
+    let out = split_lyrics_lines(vec![ll(0, 1000, "short line")], SplitConfig::default());
+    assert_eq!(out.len(), 1);
+    assert_eq!(out[0].en, "short line");
+}
+
+#[test]
+fn lyrics_long_line_splits_to_max_chars() {
+    // "His name will bring complete breakthrough" = 41 chars > 32 → splits.
+    let out = split_lyrics_lines(
+        vec![ll(0, 4000, "His name will bring complete breakthrough")],
+        SplitConfig::default(),
+    );
+    assert!(out.len() >= 2, "long line must split");
+    assert!(
+        out.iter().all(|l| l.en.chars().count() <= 32),
+        "every output line ≤ 32 chars: {:?}",
+        out.iter().map(|l| l.en.clone()).collect::<Vec<_>>()
+    );
+    // Timing covers the original span, monotonic, no synthesis beyond the split point.
+    assert_eq!(out[0].start_ms, 0);
+    assert_eq!(out.last().unwrap().end_ms, 4000);
+    assert!(
+        out[0].end_ms > 0 && out[0].end_ms < 4000,
+        "proportional mid split"
+    );
+    assert!(out.iter().all(|l| l.words.is_none()));
+}
+
+#[test]
+fn lyrics_split_timing_is_proportional_not_uniform() {
+    // Left half much longer than right → left gets proportionally more time
+    // (NOT a 50/50 uniform split). Per feedback_no_even_distribution.
+    let out = split_lyrics_lines(
+        vec![ll(0, 1000, "averylongleadingword tiny")],
+        SplitConfig { max_chars: 20 },
+    );
+    assert_eq!(out.len(), 2);
+    // left "averylongleadingword" (20 non-ws chars) vs right "tiny" (4) →
+    // mid ≈ 1000*20/24 ≈ 833, far from the uniform 500.
+    assert!(
+        out[0].end_ms > 700,
+        "left half gets most of the time, got {}",
+        out[0].end_ms
+    );
+}
